fix(frontend): use `this.fetch` instead of `fetch` in `updateTraditionMetadata`
@jorisvanzundert The issue was caused by the fact that
plain `fetch` was used instead of `this.fetch`. The former
refers to the JS builtin, the latter is the custom function
I created for services in which `baseFetch` is called
with automatic endpoint transformation. Using the plain `fetch`
call, the endpoint was not transformed properly, hence it was
forwarded to the middleware in an unexpected format, therefore
the route pattern-matching logic got tricked and
the route with the `stemmata`suffix got invoked.

TLDR: please use `this.fetch` instead of `fetch` inside
service classes.